src_lambda_pdf_to_html/requirements.txt
pdf2image==1.17.0
amazon-textract-textractor==1.8.3
